Time: 165 ms (71.84%), Space: 35 MB (98.47%) - LeetHub

+++ 1383-maximum-performance-of-a-team/1383-maximum-performance-of-a-team.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     int maxPerformance(int n, vector<int>& speed, vector<int>& efficiency, int k) {
-        priority_queue<long long, vector<long long>, greater<long long>> pq;
+        priority_queue<int, vector<int>, greater<int>> pq;
         vector<int> idx(n);
         iota(idx.begin(), idx.end(), 0);
         sort(idx.begin(), idx.end(), [&](const int &a, const int &b){
             return efficiency[a] > efficiency[b];
         });
-        long long ans = 0, sum = 0;
+        long ans = 0, sum = 0;
         for(int i = 0; i < n; i++)
         {
             pq.push(speed[idx[i]]);
